feat(tools/quantize): support .ini format
feat(src/layer): add mha int8
