+++ docs/developer-guide/operators.md
@@ -1084,9 +1084,10 @@ y = affine(out)
 | 0         | embed_dim     | int   | 0         |                   |
 | 1         | num_head      | int   | 1         |                   |
 | 2         | weight_data_size| int | 0         |                   |
+| 3         | int8_scale_term| int | 0         |                   |
 
-| weight        | type  | shape                 |
-| ------------- | ----- | --------------------- |
+| weight        | type  | shape                 | description |
+| ------------- | ----- | --- | --------------------- |
 | q_weight_data | float/fp16/int8 | [weight_data_size] |
 | q_bias_data   | float | [embed_dim]           |
 | k_weight_data | float/fp16/int8 | [weight_data_size] |
@@ -1095,6 +1096,14 @@ y = affine(out)
 | v_bias_data   | float | [embed_dim]           |
 | out_weight_data| float/fp16/int8 | [weight_data_size] |
 | out_bias_data | float | [embed_dim]           |
+| q_input_scale | float | [1] |
+| k_input_scale | float | [1] |
+| v_input_scale | float | [1] |
+| q_weight_scales | float | [embed_dim] |
+| k_weight_scales | float | [embed_dim] |
+| v_weight_scales | float | [embed_dim] |
+| internal_scales | float | [5] | scales for xq/xk/xv/before_softmax/before_output |
+
 
 # MVN
 ```
