+++ src/layer/multiheadattention.h
@@ -30,10 +30,15 @@ public:
 
     virtual int forward(const std::vector<Mat>& bottom_blobs, std::vector<Mat>& top_blobs, const Option& opt) const;
 
+#ifdef NCNN_INT8
+    int forward_int8(const std::vector<Mat>& bottom_blobs, std::vector<Mat>& top_blobs, const Option& opt) const;
+#endif
+
 public:
     int embed_dim;
     int num_head;
     int weight_data_size;
+    int int8_scale_term;
 
     Mat q_weight_data;
     Mat q_bias_data;
@@ -43,6 +48,31 @@ public:
     Mat v_bias_data;
     Mat out_weight_data;
     Mat out_bias_data;
+
+#ifdef NCNN_INT8
+    Mat q_input_scale;
+    Mat k_input_scale;
+    Mat v_input_scale;
+
+    Mat q_weight_scales;
+    Mat k_weight_scales;
+    Mat v_weight_scales;
+    Mat o_weight_scales;
+
+    /**
+     * @brief mha consists of multiple GEMM, they also have input scale
+     *
+     * internal_scales = [
+     *  q_affine_scale,
+     *  k_affine_scale,
+     *  v_affine_scale,
+     *  energy_scale,
+     *  feat_scael
+     * ]
+     *
+     */
+    Mat internal_scales;
+#endif
 };
 
 } // namespace ncnn
