+++ tools/quantize/imreadwrite.h
@@ -11,7 +11,7 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
-
+#pragma once
 #ifndef IMREADWRITE_H
 #define IMREADWRITE_H
